Fixed issue with taskList scope being mismanaged
